Moving code that doesn't update state to EditboxHelper
